Add interface for test generation and move test generation logic away from Schaapi class
