+++ modules/common/src/main/kotlin/org/cafejojo/schaapi/common/Pattern.kt
@@ -1,0 +1,3 @@
+package org.cafejojo.schaapi.common
+
+typealias Pattern = List<Node>
