+++ src/main/kotlin/org/cafejojo/schaapi/Schaapi.kt
@@ -14,11 +14,9 @@ import org.cafejojo.schaapi.patternfilter.jimple.LengthPatternFilter
 import org.cafejojo.schaapi.projectcompiler.javamaven.JavaMavenProject
 import org.cafejojo.schaapi.projectcompiler.javamaven.MavenInstaller
 import org.cafejojo.schaapi.testgenerator.jimpleevosuite.TestGenerator
-import org.cafejojo.schaapi.testgenerator.jimpleevosuite.TestableGenerator
 import org.cafejojo.schaapi.usagegraphgenerator.jimple.LibraryUsageGraphGenerator
 import java.io.File
 
-private const val DEFAULT_PATTERN_CLASS_NAME = "RegressionTest"
 private const val DEFAULT_TEST_GENERATOR_TIMEOUT = "60"
 private const val DEFAULT_PATTERN_DETECTOR_MINIMUM_COUNT = "3"
 
@@ -34,8 +32,6 @@ fun main(args: Array<String>) {
 
     val mavenDir = File(cmd.getOptionValue("maven_dir") ?: MavenInstaller.DEFAULT_MAVEN_HOME.absolutePath)
     val output = File(cmd.getOptionValue('o')).apply { mkdirs() }
-    val outputPatterns = output.resolve("patterns/").apply { mkdirs() }
-    val outputTests = output.resolve("tests/").apply { mkdirs() }
     val library = JavaMavenProject(File(cmd.getOptionValue('l')), mavenDir)
     val users = cmd.getOptionValues('u').map { JavaMavenProject(File(it), mavenDir) }
 
@@ -57,23 +53,16 @@ fun main(args: Array<String>) {
         .filter { IncompleteInitPatternFilter().retain(it) }
         .filter { LengthPatternFilter().retain(it) }
 
-    val classGenerator = TestableGenerator(DEFAULT_PATTERN_CLASS_NAME)
-    patterns.forEachIndexed { index, pattern ->
-        classGenerator.generateMethod("pattern$index", pattern)
-    }
-
-    classGenerator.writeToFile(outputPatterns.absolutePath)
-
     val testGeneratorTimeout = cmd.getOptionOrDefault("test_generator_timeout", DEFAULT_TEST_GENERATOR_TIMEOUT).toInt()
     val testGeneratorEnableOutput = cmd.hasOption("test_generator_enable_output")
+
     TestGenerator(
-        fullyQualifiedClassName = DEFAULT_PATTERN_CLASS_NAME,
-        classpath = outputPatterns.absolutePath + File.pathSeparator + library.classpath,
-        outputDirectory = outputTests.absolutePath,
-        generationTimeoutSeconds = testGeneratorTimeout,
+        library = library,
+        outputDirectory = output,
+        timeout = testGeneratorTimeout,
         processStandardStream = if (testGeneratorEnableOutput) System.out else null,
         processErrorStream = if (testGeneratorEnableOutput) System.out else null
-    ).run()
+    ).generate(patterns)
 }
 
 private fun buildOptions(): Options =
